fix crop anchor
