--- build.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/bash
-
-go build -ldflags '-w -s' .
