+++ config.toml
@@ -10,4 +10,4 @@ path = 'output/{year}'
 ## 协议：socks4, socks5, http, https
 socket = "socks5://127.0.0.1:7891"
 ## 设置是否启用代理
-enable = true
+enable = false
