+++ go.mod
@@ -7,9 +7,12 @@ require (
 	github.com/disintegration/imaging v1.6.2
 	github.com/imroc/req v0.3.0
 	github.com/sirupsen/logrus v1.2.0
+	github.com/spf13/afero v1.2.2 // indirect
 	github.com/spf13/viper v1.6.2
 	github.com/stretchr/testify v1.4.0 // indirect
 	golang.org/x/net v0.0.0-20201021035429-f5854403a974
-	golang.org/x/sys v0.0.0-20210220050731-9a76102bfb43 // indirect
+	golang.org/x/sys v0.0.0-20210326220804-49726bf1d181 // indirect
 	golang.org/x/text v0.3.3
+	gopkg.in/check.v1 v1.0.0-20190902080502-41f04d3bba15 // indirect
+	gopkg.in/yaml.v2 v2.2.8 // indirect
 )
