+++ pkg/scraper/dmm_test.go
@@ -1,8 +1,49 @@
 package scraper
 
 import (
+	myclient "better-av-tool/pkg/client"
+	"better-av-tool/pkg/logger"
+	"github.com/PuerkitoBio/goquery"
 	"testing"
 )
+
+var dmmTests map[string]*DMMScraper
+var fc2Tests map[string]*Fc2Scraper
+var mgsTests map[string]*MGStageScraper
+var heyzoTests map[string]*HeyzoScraper
+var gyuttoTests map[string]*GyuttoScraper
+
+type fields struct {
+	doc *goquery.Document
+}
+
+type args struct {
+	query string
+	url   string
+}
+
+type testCase struct {
+	name    string
+	fields  fields
+	args    args
+	wantErr bool
+	want    string
+}
+
+func Setup() {
+	fc2Tests = make(map[string]*Fc2Scraper)
+	mgsTests = make(map[string]*MGStageScraper)
+	heyzoTests = make(map[string]*HeyzoScraper)
+	gyuttoTests = make(map[string]*GyuttoScraper)
+	dmmTests = make(map[string]*DMMScraper)
+
+	log = logger.New()
+
+	client = myclient.New()
+	client.SetProxyUrl("socks5://192.168.0.110:7891")
+
+	Init(client, log)
+}
 
 func TestDMMScraper_FetchDoc(t *testing.T) {
 	tests := []testCase{
