+++ pkg/scraper/scraper.go
@@ -1,12 +1,36 @@
 package scraper
 
 import (
-	myclient "better-av-tool/internal/client"
-	"better-av-tool/internal/logger"
-	"net/http"
-
+	myclient "better-av-tool/pkg/client"
+	"better-av-tool/pkg/logger"
 	"github.com/PuerkitoBio/goquery"
+	"net/http"
 )
+
+// Scraper is interface
+type Scraper interface {
+	// Remote
+	FetchDoc(query string) (err error)
+
+	// Local
+	GetPlot() string
+	GetTitle() string
+	GetDirector() string
+	GetRuntime() string
+	GetTags() []string
+	GetMaker() string
+	GetActors() []string
+	GetLabel() string
+	GetNumber() string
+	GetCover() string
+	GetWebsite() string
+	GetPremiered() string
+	GetYear() string
+	GetSeries() string
+
+	// Operation
+	NeedCut() bool
+}
 
 var (
 	client myclient.Client
