+++ tests/testcase.sh
@@ -1,0 +1,6 @@
+touch 'h_479gne218.mp4'
+touch 'sdde-552.mp4'
+touch 'FC2-1027251.mp4'
+touch 'heyzo-1031.mp4'
+touch 'siro-3171.mp4'
+./better-av-tool.exe
